divide app into screens (still not functional) - notes provider structure
