+++ pubspec.yaml
@@ -24,6 +24,8 @@ dependencies:
   flutter_local_notifications: any
   gradient_app_bar: any
   shared_preferences: any
+  path_provider: ^1.5.1
+  sembast: ^2.1.2+2
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
